+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * `aoc`: Fixed `find` command not working. (undefined variable)
   * Breaking Changes:
     * Internal: Basic REST calls now return data directly. (no more `data` key). For advanced calls, use `call`.
+    * Internal: Transfer SDK download is now a 2-step procedure: First get the YAML file from github with URLS for the various platforms and versions, and then download the archive from the official IBM repository.
   
 * 4.19.0
   * New Features: